Remove more dead code from unets implementation
