+++ kornia/feature/disk/unets/unet.py
@@ -3,19 +3,6 @@ import torch.nn as nn
 from kornia.core import Tensor
 
 from .blocks import ThinUnetDownBlock, ThinUnetUpBlock
-from .ops import NoOp, TrivialDownsample, TrivialUpsample
-
-thin_setup = {
-    'gate': nn.PReLU,
-    'norm': nn.InstanceNorm2d,
-    'upsample': TrivialUpsample,
-    'downsample': TrivialDownsample,
-    'down_block': ThinUnetDownBlock,
-    'up_block': ThinUnetUpBlock,
-    'dropout': NoOp,
-    'padding': True,
-    'bias': True,
-}
 
 
 class Unet(nn.Module):
@@ -29,20 +16,17 @@ class Unet(nn.Module):
         self.down = down
         self.in_features = in_features
 
-        DownBlock = thin_setup['down_block']
-        UpBlock = thin_setup['up_block']
-
         down_dims = [in_features] + down
         self.path_down = nn.ModuleList()
         for i, (d_in, d_out) in enumerate(zip(down_dims[:-1], down_dims[1:])):
-            block = DownBlock(d_in, d_out, size=size, name=f'down_{i}', setup=thin_setup, is_first=i == 0)
+            block = ThinUnetDownBlock(d_in, d_out, size=size, is_first=i == 0)
             self.path_down.append(block)
 
         bot_dims = [down[-1]] + up
         hor_dims = down_dims[-2::-1]
         self.path_up = nn.ModuleList()
         for i, (d_bot, d_hor, d_out) in enumerate(zip(bot_dims, hor_dims, up)):
-            block = UpBlock(d_bot, d_hor, d_out, size=size, name=f'up_{i}', setup=thin_setup)
+            block = ThinUnetUpBlock(d_bot, d_hor, d_out, size=size)
             self.path_up.append(block)
 
         self.n_params = 0
